refactored old triangle/pentagonal/hexagonal number classes to use new general PolygonalNumber class

+++ src/test/scala/com/msokoryansky/MathUtils/TestHexagonalNumber.scala
@@ -49,7 +49,7 @@ class TestHexagonalNumber extends FunSuite {
     val mismatches = for {
       i <- 1 to 10000000
       index = HexagonalNumber.getHexagonalNumberIndex(HexagonalNumber.hexagonalNumber(i)).getOrElse(0)
-      if i != index
+      if i.toLong != index
     } yield (i, index)
     assert(mismatches.isEmpty)
   }
