+++ src/test/scala/com/msokoryansky/MathUtils/TestTriangleNumber.scala
@@ -45,7 +45,7 @@ class TestTriangleNumber extends FunSuite {
     val mismatches = for {
       i <- 1 to 10000000
       index = TriangleNumber.getTriangleNumberIndex(TriangleNumber.triangleNumber(i)).getOrElse(0)
-      if i != index
+      if i.toLong != index
     } yield (i, index)
     assert(mismatches.isEmpty)
   }
